Don't use production names when compiling plist
Building these sources requires combining master and instance binary
tables, which for now means that glyph names must be consistent
between fonts. At least we can use development names to make debugging
easier, and perhaps at some point this won't be an issue (if some
workaround is implemented in FontTools) so we can take out these flags
entirely.

The alternative is to try use the instances' naming settings when
compiling the masters, but fontmake isn't really set up to do that and
it would be a bit complicated for something that might be fixed later
anyways.