Update Armenian version numbers for release.
Previously these were 2.900+, because the phase 2 versions had major
versions of 2.

This seems like a good time to unify the major version numbering to be
consistent across the noto phase, so we will keep the major version at
2 but bump the minor.  The minor of 040 was chosen so there's a bump
in the second digit, since we're also converting to a three-digit minor
with this phase of Noto.