Normalize .glyphs files by reading/writing them using glyphsLib.
This is in preparation for swatting the file names, so that the name
changes will be uncluttered by extraneous changes caused by reading/
writing the .glyphs files.

Glyphs quotes .appVersion, panose values, and weight values, but
glyphsLib treats panose and weight values as ints.  It also treats
appVersion as an int.  Unfortunately treating it as a string still
results in it being unquoted, because in many places Glyphs doesn't
quote symbols that don't need it, and numeric appVersion strings
follow this pattern-- Glyphs however appears to treat appVersion
specially.

glyphsLib represents the .appVersion as an unquoted string, but the
glyphs files quote it.