Add Noto Sans Syriac
This file contains all of Estrangela, Eastern and Western Syriac in a
single font. Applications can choose the preferred style by requesting
the corresponding OpenType language system. We’ll need this for
re-packaging Noto Sans into a single pan-Unicode font file.

The OpenType language system tags for Syriac have been registered with
Microsoft and became part of the OpenType 1.8 specification; HarfBuzz
recognizes the BCP47 script codes for Syriac and converts them to the
correct OpenType language system codes.

The font design is probably not the final version, but I’d like to
already add the current versino to the source repository so that
applications and operating systems can be tested early.