fix familyName of NotoSans Old North Arabian and Sora Sompeng
These source files are the only two for which the `familyName` attribute
uses a shortened form that doesn't match the respective file name.

This patch sets the familyName to the long (preferred) name, and sets
the `styleMapFamilyName` to the short name.

It also removes the `preferredFamilyName`, `fileName` and
`postscriptFontName` custom parameters since they are now redundant,
as they are the same as the default values.

Fixes #107